get correct date for all views in calendar to retrieve events for iu sports
